# backend/gen/http/notes/server/encode_decode.go
// Code generated by goa, DO NOT EDIT.
//
// notes HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/flaccid/google-keep-clone/backend/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strconv"

	notes "github.com/flaccid/google-keep-clone/backend/gen/notes"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCreateResponse returns an encoder for responses returned by the notes
// create endpoint.
func EncodeCreateResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewCreateResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeCreateRequest returns a decoder for requests sent to the notes create
// endpoint.
func DecodeCreateRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.CreatePayload, error) {
	return func(r *http.Request) (*notes.CreatePayload, error) {
		var payload *notes.CreatePayload
		var (
			body CreateRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				err = nil
			} else {
				var gerr *goa.ServiceError
				if errors.As(err, &gerr) {
					return payload, gerr
				}
				return payload, goa.DecodePayloadError(err.Error())
			}
		}
		err = ValidateCreateRequestBody(&body)
		if err != nil {
			return payload, err
		}
		payload = NewCreatePayload(&body)

		return payload, nil
	}
}

// EncodeGetResponse returns an encoder for responses returned by the notes get
// endpoint.
func EncodeGetResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewGetResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetRequest returns a decoder for requests sent to the notes get
// endpoint.
func DecodeGetRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.GetPayload, error) {
	return func(r *http.Request) (*notes.GetPayload, error) {
		var payload *notes.GetPayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewGetPayload(id)

		return payload, nil
	}
}

// EncodeListResponse returns an encoder for responses returned by the notes
// list endpoint.
func EncodeListResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.ListNotesResponse)
		enc := encoder(ctx, w)
		body := NewListResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListRequest returns a decoder for requests sent to the notes list
// endpoint.
func DecodeListRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.ListPayload, error) {
	return func(r *http.Request) (*notes.ListPayload, error) {
		var payload *notes.ListPayload
		var (
			pageSize  *int
			pageToken *string
			filter    *string
			err       error
		)
		qp := r.URL.Query()
		{
			pageSizeRaw := qp.Get("pageSize")
			if pageSizeRaw != "" {
				v, err2 := strconv.ParseInt(pageSizeRaw, 10, strconv.IntSize)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("pageSize", pageSizeRaw, "integer"))
				}
				pv := int(v)
				pageSize = &pv
			}
		}
		pageTokenRaw := qp.Get("pageToken")
		if pageTokenRaw != "" {
			pageToken = &pageTokenRaw
		}
		filterRaw := qp.Get("filter")
		if filterRaw != "" {
			filter = &filterRaw
		}
		if err != nil {
			return payload, err
		}
		payload = NewListPayload(pageSize, pageToken, filter)

		return payload, nil
	}
}

// EncodeUpdateResponse returns an encoder for responses returned by the notes
// update endpoint.
func EncodeUpdateResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewUpdateResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeUpdateRequest returns a decoder for requests sent to the notes update
// endpoint.
func DecodeUpdateRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.UpdatePayload, error) {
	return func(r *http.Request) (*notes.UpdatePayload, error) {
		var payload *notes.UpdatePayload
		var (
			body UpdateRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return payload, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return payload, gerr
			}
			return payload, goa.DecodePayloadError(err.Error())
		}
		err = ValidateUpdateRequestBody(&body)
		if err != nil {
			return payload, err
		}

		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewUpdatePayload(&body, id)

		return payload, nil
	}
}

// EncodeDeleteResponse returns an encoder for responses returned by the notes
// delete endpoint.
func EncodeDeleteResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// DecodeDeleteRequest returns a decoder for requests sent to the notes delete
// endpoint.
func DecodeDeleteRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.DeletePayload, error) {
	return func(r *http.Request) (*notes.DeletePayload, error) {
		var payload *notes.DeletePayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewDeletePayload(id)

		return payload, nil
	}
}

// EncodePinResponse returns an encoder for responses returned by the notes pin
// endpoint.
func EncodePinResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewPinResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodePinRequest returns a decoder for requests sent to the notes pin
// endpoint.
func DecodePinRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.PinPayload, error) {
	return func(r *http.Request) (*notes.PinPayload, error) {
		var payload *notes.PinPayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewPinPayload(id)

		return payload, nil
	}
}

// EncodeUnpinResponse returns an encoder for responses returned by the notes
// unpin endpoint.
func EncodeUnpinResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewUnpinResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeUnpinRequest returns a decoder for requests sent to the notes unpin
// endpoint.
func DecodeUnpinRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.UnpinPayload, error) {
	return func(r *http.Request) (*notes.UnpinPayload, error) {
		var payload *notes.UnpinPayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewUnpinPayload(id)

		return payload, nil
	}
}

// EncodeArchiveResponse returns an encoder for responses returned by the notes
// archive endpoint.
func EncodeArchiveResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewArchiveResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeArchiveRequest returns a decoder for requests sent to the notes
// archive endpoint.
func DecodeArchiveRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.ArchivePayload, error) {
	return func(r *http.Request) (*notes.ArchivePayload, error) {
		var payload *notes.ArchivePayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewArchivePayload(id)

		return payload, nil
	}
}

// EncodeUnarchiveResponse returns an encoder for responses returned by the
// notes unarchive endpoint.
func EncodeUnarchiveResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewUnarchiveResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeUnarchiveRequest returns a decoder for requests sent to the notes
// unarchive endpoint.
func DecodeUnarchiveRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.UnarchivePayload, error) {
	return func(r *http.Request) (*notes.UnarchivePayload, error) {
		var payload *notes.UnarchivePayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewUnarchivePayload(id)

		return payload, nil
	}
}

// EncodeTrashResponse returns an encoder for responses returned by the notes
// trash endpoint.
func EncodeTrashResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewTrashResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeTrashRequest returns a decoder for requests sent to the notes trash
// endpoint.
func DecodeTrashRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.TrashPayload, error) {
	return func(r *http.Request) (*notes.TrashPayload, error) {
		var payload *notes.TrashPayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewTrashPayload(id)

		return payload, nil
	}
}

// EncodeRestoreResponse returns an encoder for responses returned by the notes
// restore endpoint.
func EncodeRestoreResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*notes.Note)
		enc := encoder(ctx, w)
		body := NewRestoreResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeRestoreRequest returns a decoder for requests sent to the notes
// restore endpoint.
func DecodeRestoreRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*notes.RestorePayload, error) {
	return func(r *http.Request) (*notes.RestorePayload, error) {
		var payload *notes.RestorePayload
		var (
			id string

			params = mux.Vars(r)
		)
		id = params["id"]
		payload = NewRestorePayload(id)

		return payload, nil
	}
}

// unmarshalSectionRequestBodyRequestBodyToNotesSection builds a value of type
// *notes.Section from a value of type *SectionRequestBodyRequestBody.
func unmarshalSectionRequestBodyRequestBodyToNotesSection(v *SectionRequestBodyRequestBody) *notes.Section {
	if v == nil {
		return nil
	}
	res := &notes.Section{}
	if v.Text != nil {
		res.Text = unmarshalTextContentRequestBodyRequestBodyToNotesTextContent(v.Text)
	}
	if v.List != nil {
		res.List = unmarshalListContentRequestBodyRequestBodyToNotesListContent(v.List)
	}

	return res
}

// unmarshalTextContentRequestBodyRequestBodyToNotesTextContent builds a value
// of type *notes.TextContent from a value of type
// *TextContentRequestBodyRequestBody.
func unmarshalTextContentRequestBodyRequestBodyToNotesTextContent(v *TextContentRequestBodyRequestBody) *notes.TextContent {
	if v == nil {
		return nil
	}
	res := &notes.TextContent{
		Text: v.Text,
	}

	return res
}

// unmarshalListContentRequestBodyRequestBodyToNotesListContent builds a value
// of type *notes.ListContent from a value of type
// *ListContentRequestBodyRequestBody.
func unmarshalListContentRequestBodyRequestBodyToNotesListContent(v *ListContentRequestBodyRequestBody) *notes.ListContent {
	if v == nil {
		return nil
	}
	res := &notes.ListContent{}
	if v.ListItems != nil {
		res.ListItems = make([]*notes.ListItem, len(v.ListItems))
		for i, val := range v.ListItems {
			if val == nil {
				res.ListItems[i] = nil
				continue
			}
			res.ListItems[i] = unmarshalListItemRequestBodyRequestBodyToNotesListItem(val)
		}
	}

	return res
}

// unmarshalListItemRequestBodyRequestBodyToNotesListItem builds a value of
// type *notes.ListItem from a value of type *ListItemRequestBodyRequestBody.
func unmarshalListItemRequestBodyRequestBodyToNotesListItem(v *ListItemRequestBodyRequestBody) *notes.ListItem {
	if v == nil {
		return nil
	}
	res := &notes.ListItem{
		Checked: v.Checked,
	}
	if v.Text != nil {
		res.Text = unmarshalTextContentRequestBodyRequestBodyToNotesTextContent(v.Text)
	}
	if v.ChildListItems != nil {
		res.ChildListItems = make([]*notes.ListItem, len(v.ChildListItems))
		for i, val := range v.ChildListItems {
			if val == nil {
				res.ChildListItems[i] = nil
				continue
			}
			res.ChildListItems[i] = unmarshalListItemRequestBodyRequestBodyToNotesListItem(val)
		}
	}

	return res
}

// marshalNotesSectionToSectionResponseBody builds a value of type
// *SectionResponseBody from a value of type *notes.Section.
func marshalNotesSectionToSectionResponseBody(v *notes.Section) *SectionResponseBody {
	if v == nil {
		return nil
	}
	res := &SectionResponseBody{}
	if v.Text != nil {
		res.Text = marshalNotesTextContentToTextContentResponseBody(v.Text)
	}
	if v.List != nil {
		res.List = marshalNotesListContentToListContentResponseBody(v.List)
	}

	return res
}

// marshalNotesTextContentToTextContentResponseBody builds a value of type
// *TextContentResponseBody from a value of type *notes.TextContent.
func marshalNotesTextContentToTextContentResponseBody(v *notes.TextContent) *TextContentResponseBody {
	if v == nil {
		return nil
	}
	res := &TextContentResponseBody{
		Text: v.Text,
	}

	return res
}

// marshalNotesListContentToListContentResponseBody builds a value of type
// *ListContentResponseBody from a value of type *notes.ListContent.
func marshalNotesListContentToListContentResponseBody(v *notes.ListContent) *ListContentResponseBody {
	if v == nil {
		return nil
	}
	res := &ListContentResponseBody{}
	if v.ListItems != nil {
		res.ListItems = make([]*ListItemResponseBody, len(v.ListItems))
		for i, val := range v.ListItems {
			if val == nil {
				res.ListItems[i] = nil
				continue
			}
			res.ListItems[i] = marshalNotesListItemToListItemResponseBody(val)
		}
	}

	return res
}

// marshalNotesListItemToListItemResponseBody builds a value of type
// *ListItemResponseBody from a value of type *notes.ListItem.
func marshalNotesListItemToListItemResponseBody(v *notes.ListItem) *ListItemResponseBody {
	if v == nil {
		return nil
	}
	res := &ListItemResponseBody{
		Checked: v.Checked,
	}
	if v.Text != nil {
		res.Text = marshalNotesTextContentToTextContentResponseBody(v.Text)
	}
	if v.ChildListItems != nil {
		res.ChildListItems = make([]*ListItemResponseBody, len(v.ChildListItems))
		for i, val := range v.ChildListItems {
			if val == nil {
				res.ChildListItems[i] = nil
				continue
			}
			res.ChildListItems[i] = marshalNotesListItemToListItemResponseBody(val)
		}
	}

	return res
}

// marshalNotesAttachmentToAttachmentResponseBody builds a value of type
// *AttachmentResponseBody from a value of type *notes.Attachment.
func marshalNotesAttachmentToAttachmentResponseBody(v *notes.Attachment) *AttachmentResponseBody {
	if v == nil {
		return nil
	}
	res := &AttachmentResponseBody{
		Name: v.Name,
	}
	if v.MimeType != nil {
		res.MimeType = make([]string, len(v.MimeType))
		for i, val := range v.MimeType {
			res.MimeType[i] = val
		}
	}

	return res
}

// marshalNotesPermissionToPermissionResponseBody builds a value of type
// *PermissionResponseBody from a value of type *notes.Permission.
func marshalNotesPermissionToPermissionResponseBody(v *notes.Permission) *PermissionResponseBody {
	if v == nil {
		return nil
	}
	res := &PermissionResponseBody{
		Name:    v.Name,
		Email:   v.Email,
		Deleted: v.Deleted,
	}
	if v.Role != nil {
		role := string(*v.Role)
		res.Role = &role
	}
	if v.User != nil {
		res.User = marshalNotesUserToUserResponseBody(v.User)
	}
	if v.Group != nil {
		res.Group = marshalNotesGroupToGroupResponseBody(v.Group)
	}
	if v.Family != nil {
		res.Family = marshalNotesFamilyToFamilyResponseBody(v.Family)
	}

	return res
}

// marshalNotesUserToUserResponseBody builds a value of type *UserResponseBody
// from a value of type *notes.User.
func marshalNotesUserToUserResponseBody(v *notes.User) *UserResponseBody {
	if v == nil {
		return nil
	}
	res := &UserResponseBody{
		Email: v.Email,
	}

	return res
}

// marshalNotesGroupToGroupResponseBody builds a value of type
// *GroupResponseBody from a value of type *notes.Group.
func marshalNotesGroupToGroupResponseBody(v *notes.Group) *GroupResponseBody {
	if v == nil {
		return nil
	}
	res := &GroupResponseBody{
		Email: v.Email,
	}

	return res
}

// marshalNotesFamilyToFamilyResponseBody builds a value of type
// *FamilyResponseBody from a value of type *notes.Family.
func marshalNotesFamilyToFamilyResponseBody(v *notes.Family) *FamilyResponseBody {
	if v == nil {
		return nil
	}
	res := &FamilyResponseBody{}

	return res
}

// marshalNotesNoteToNoteResponseBody builds a value of type *NoteResponseBody
// from a value of type *notes.Note.
func marshalNotesNoteToNoteResponseBody(v *notes.Note) *NoteResponseBody {
	if v == nil {
		return nil
	}
	res := &NoteResponseBody{
		Name:       v.Name,
		CreateTime: v.CreateTime,
		UpdateTime: v.UpdateTime,
		TrashTime:  v.TrashTime,
		Trashed:    v.Trashed,
		Title:      v.Title,
		Pinned:     v.Pinned,
		Archived:   v.Archived,
	}
	if v.Color != nil {
		color := string(*v.Color)
		res.Color = &color
	}
	if v.Body != nil {
		res.Body = marshalNotesSectionToSectionResponseBody(v.Body)
	}
	if v.Attachments != nil {
		res.Attachments = make([]*AttachmentResponseBody, len(v.Attachments))
		for i, val := range v.Attachments {
			if val == nil {
				res.Attachments[i] = nil
				continue
			}
			res.Attachments[i] = marshalNotesAttachmentToAttachmentResponseBody(val)
		}
	}
	if v.Permissions != nil {
		res.Permissions = make([]*PermissionResponseBody, len(v.Permissions))
		for i, val := range v.Permissions {
			if val == nil {
				res.Permissions[i] = nil
				continue
			}
			res.Permissions[i] = marshalNotesPermissionToPermissionResponseBody(val)
		}
	}
	if v.Labels != nil {
		res.Labels = make([]string, len(v.Labels))
		for i, val := range v.Labels {
			res.Labels[i] = val
		}
	}

	return res
}