# backend/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Google Keep Clone API
    description: A clone of the Google Keep REST API, designed to mirror the official API surface.
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /v1/notes/{id}:archive:
        post:
            tags:
                - notes
            summary: archive notes
            description: Archives a note.
            operationId: notes#archive
            parameters:
                - name: id
                  in: path
                  description: The ID of the note.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Note'
            schemes:
                - http
...
    /v1/notes/{id}:restore:
        post:
            tags:
                - notes
            summary: restore notes
            description: Restores a trashed note.
            operationId: notes#restore
            parameters:
                - name: id
                  in: path
                  description: The ID of the note.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Note'
            schemes:
                - http
...
    /v1/notes/{id}:unarchive:
        post:
            tags:
                - notes
            summary: unarchive notes
            description: Unarchives a note.
            operationId: notes#unarchive
            parameters:
                - name: id
                  in: path
                  description: The ID of the note.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Note'
            schemes:
                - http
...
definitions:
    Attachment:
        title: Attachment
        type: object
        properties:
            mimeType:
                type: array
                items:
                    type: string
                    example: Labore quis quo.
                description: The MIME types in which the attachment is available.
                example:
                    - Ut blanditiis natus voluptatem aut voluptatibus.
                    - Maxime est autem natus eum autem.
            name:
                type: string
                description: The resource name of the attachment.
                example: Ipsa voluptatibus repudiandae modi quos.
        description: An attachment to a note.
        example:
            mimeType:
                - Natus quo earum.
                - Debitis sed at.
                - Ab ex molestiae sit aut esse laboriosam.
            name: In nihil.
...
    Family:
        title: Family
        type: object
        description: Describes a single Google Family. Empty type.
        example: {}
    Group:
        title: Group
        type: object
        properties:
            email:
                type: string
                description: The group email.
                example: Consequatur officia est suscipit eaque.
        description: Describes a single group.
        example:
            email: Aperiam omnis beatae sunt quis quia.
...
    Note:
        title: Note
        type: object
        properties:
            archived:
                type: boolean
                description: Whether the note is archived.
                example: true
            attachments:
                type: array
                items:
                    $ref: '#/definitions/Attachment'
                description: Output only. The attachments attached to this note.
                example:
                    - mimeType:
                        - Quia aut modi ut voluptas.
                        - Minus vitae officia iusto.
                        - Blanditiis assumenda sit eos.
                        - Deserunt et tempora.
                      name: Voluptas pariatur velit facere nihil.
                    - mimeType:
                        - Quia aut modi ut voluptas.
                        - Minus vitae officia iusto.
                        - Blanditiis assumenda sit eos.
                        - Deserunt et tempora.
                      name: Voluptas pariatur velit facere nihil.
            body:
                $ref: '#/definitions/Section'
            color:
                type: string
                description: The color of the note.
                example: BLUE
                enum:
                    - DEFAULT
                    - RED
                    - ORANGE
                    - YELLOW
                    - GREEN
                    - TEAL
                    - BLUE
                    - CERULEAN
                    - PURPLE
                    - PINK
                    - BROWN
                    - GRAY
            createTime:
                type: string
                description: Output only. When this note was created (RFC3339 format).
                example: Est omnis.
            labels:
                type: array
                items:
                    type: string
                    example: Qui quia praesentium quae dolor saepe quam.
                description: The labels assigned to this note.
                example:
                    - Necessitatibus et neque.
                    - Qui nihil eveniet.
                    - Neque quo esse soluta earum illum.
            name:
                type: string
                description: Output only. The resource name of this note (e.g. 'notes/{uuid}').
                example: Illum maiores.
            permissions:
                type: array
                items:
                    $ref: '#/definitions/Permission'
                description: Output only. The list of permissions set on the note.
                example:
                    - deleted: true
                      email: Ut debitis velit numquam.
                      family: {}
                      group:
                        email: Est quas.
                      name: Est quis saepe velit nemo consequatur.
                      role: WRITER
                      user:
                        email: Nam quam repellendus tempore.
                    - deleted: true
                      email: Ut debitis velit numquam.
                      family: {}
                      group:
                        email: Est quas.
                      name: Est quis saepe velit nemo consequatur.
                      role: WRITER
                      user:
                        email: Nam quam repellendus tempore.
                    - deleted: true
                      email: Ut debitis velit numquam.
                      family: {}
                      group:
                        email: Est quas.
                      name: Est quis saepe velit nemo consequatur.
                      role: WRITER
                      user:
                        email: Nam quam repellendus tempore.
                    - deleted: true
                      email: Ut debitis velit numquam.
                      family: {}
                      group:
                        email: Est quas.
                      name: Est quis saepe velit nemo consequatur.
                      role: WRITER
                      user:
                        email: Nam quam repellendus tempore.
            pinned:
                type: boolean
                description: Whether the note is pinned.
                example: true
            title:
                type: string
                description: The title of the note. Length must be less than 1,000 characters.
                example: Eligendi velit quaerat temporibus.
            trashTime:
                type: string
                description: Output only. When this note was trashed. Only set if trashed.
                example: Nostrum ut ut dolorum.
            trashed:
                type: boolean
                description: Output only. True if this note has been trashed.
                example: false
            updateTime:
                type: string
                description: Output only. When this note was last modified (RFC3339 format).
                example: Dolores officia aut.
        description: The newly created note.
        example:
            archived: false
            attachments:
                - mimeType:
                    - Quia aut modi ut voluptas.
                    - Minus vitae officia iusto.
                    - Blanditiis assumenda sit eos.
                    - Deserunt et tempora.
                  name: Voluptas pariatur velit facere nihil.
                - mimeType:
                    - Quia aut modi ut voluptas.
                    - Minus vitae officia iusto.
                    - Blanditiis assumenda sit eos.
                    - Deserunt et tempora.
                  name: Voluptas pariatur velit facere nihil.
                - mimeType:
                    - Quia aut modi ut voluptas.
                    - Minus vitae officia iusto.
                    - Blanditiis assumenda sit eos.
                    - Deserunt et tempora.
                  name: Voluptas pariatur velit facere nihil.
            body:
                list:
                    listItems:
                        - checked: true
                          childListItems:
                            - {}
                            - {}
                            - {}
                            - {}
                          text:
                            text: Accusamus quod itaque.
                        - checked: true
                          childListItems:
                            - {}
                            - {}
                            - {}
                            - {}
                          text:
                            text: Accusamus quod itaque.
                        - checked: true
                          childListItems:
                            - {}
                            - {}
                            - {}
                            - {}
                          text:
                            text: Accusamus quod itaque.
                text:
                    text: Accusamus quod itaque.
            color: DEFAULT
            createTime: Similique harum accusantium sequi iusto necessitatibus accusamus.
            labels:
                - Corrupti eligendi sit provident voluptatem.
                - Autem sit.
                - Debitis ipsum et commodi deserunt alias numquam.
            name: Consequuntur quam.
            permissions:
                - deleted: true
                  email: Ut debitis velit numquam.
                  family: {}
                  group:
                    email: Est quas.
                  name: Est quis saepe velit nemo consequatur.
                  role: WRITER
                  user:
                    email: Nam quam repellendus tempore.
                - deleted: true
                  email: Ut debitis velit numquam.
                  family: {}
                  group:
                    email: Est quas.
                  name: Est quis saepe velit nemo consequatur.
                  role: WRITER
                  user:
                    email: Nam quam repellendus tempore.
                - deleted: true
                  email: Ut debitis velit numquam.
                  family: {}
                  group:
                    email: Est quas.
                  name: Est quis saepe velit nemo consequatur.
                  role: WRITER
                  user:
                    email: Nam quam repellendus tempore.
                - deleted: true
                  email: Ut debitis velit numquam.
                  family: {}
                  group:
                    email: Est quas.
                  name: Est quis saepe velit nemo consequatur.
                  role: WRITER
                  user:
                    email: Nam quam repellendus tempore.
            pinned: false
            title: Ea aliquam molestiae.
            trashTime: Eos eius ut est aliquid ratione dolor.
            trashed: true
            updateTime: Rerum quae aspernatur.
...
    Permission:
        title: Permission
        type: object
        properties:
            deleted:
                type: boolean
                description: Output only. Whether this member has been deleted.
                example: false
            email:
                type: string
                description: The email associated with the member.
                example: Quod est ipsum consequuntur totam et sit.
            family:
                $ref: '#/definitions/Family'
            group:
                $ref: '#/definitions/Group'
            name:
                type: string
                description: Output only. The resource name.
                example: Nam aut.
            role:
                type: string
                description: The role granted by this permission.
                example: WRITER
                enum:
                    - OWNER
                    - WRITER
            user:
                $ref: '#/definitions/User'
        description: A single permission on the note. Associates a member with a role.
        example:
            deleted: true
            email: Aut illum numquam.
            family: {}
            group:
                email: Est quas.
            name: Et sed est iusto est.
            role: WRITER
            user:
                email: Nam quam repellendus tempore.
    Section:
        title: Section
        type: object
        properties:
            list:
                $ref: '#/definitions/ListContent'
            text:
                $ref: '#/definitions/TextContent'
        description: The content of the note. Either text or list content.
        example:
            list:
                listItems:
                    - checked: true
                      childListItems:
                        - {}
                        - {}
                        - {}
                        - {}
                      text:
                        text: Accusamus quod itaque.
                    - checked: true
                      childListItems:
                        - {}
                        - {}
                        - {}
                        - {}
                      text:
                        text: Accusamus quod itaque.
                    - checked: true
                      childListItems:
                        - {}
                        - {}
                        - {}
                        - {}
                      text:
                        text: Accusamus quod itaque.
            text:
                text: Accusamus quod itaque.
    TextContent:
        title: TextContent
        type: object
        properties:
            text:
                type: string
                description: The text content.
                example: Et in id praesentium nemo.
        description: The block of text for a single text section or list item.
        example:
            text: Ipsa qui iure quae.
    User:
        title: User
        type: object
        properties:
            email:
                type: string
                description: The user's email.
                example: Consectetur et aperiam vitae fugiat qui explicabo.
        description: Describes a single user.
        example:
            email: Autem voluptas itaque ut.